Extract the dashboard/projects table into its own container
