+++ frontend/containers/dashboard/projects/table/index.ts
@@ -1,0 +1,2 @@
+export { default } from './component';
+export type { ProjectsTableProps } from './types';
