+++ frontend/containers/dashboard/projects/table/types.ts
@@ -1,0 +1,1 @@
+export type ProjectsTableProps = {};
